Added Unit Tests for QuestionController and service Update, delete method

+++ src/main/java/com/example/musictheory/controllers/QuestionController.java
@@ -6,6 +6,8 @@ import com.example.musictheory.services.QuizService;
 import com.example.musictheory.utils.Util;
 import jakarta.validation.Valid;
 import org.apache.commons.lang3.StringUtils;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -20,6 +22,7 @@ import java.util.Optional;
 @RequestMapping("/api/v1")
 public class QuestionController {
 
+    private static final Logger log = LogManager.getLogger(QuestionController.class);
     @Autowired
     private QuizService quizService;
 
@@ -42,9 +45,16 @@ public class QuestionController {
     @GetMapping(value = "/question")
     public ResponseEntity<Optional<Question>> getQuestion(@RequestBody String uid){
         try {
-            return new ResponseEntity<>(quizService.findQuestionByUid(uid), HttpStatus.OK);
+            if (StringUtils.isAlphanumeric(uid)) {
+                return new ResponseEntity<>(quizService.findQuestionByUid(uid), HttpStatus.OK);
+            } else{
+                throw new IOException("Invalid Uid. Retry with valid uid.");
+            }
         } catch (FileNotFoundException e) {
-            System.err.println(e);
+            log.error("e: ", e);
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        } catch (IOException e) {
+            log.error("e: ", e);
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
     }
@@ -58,7 +68,7 @@ public class QuestionController {
             return new ResponseEntity<>(quizService.createQuestion(question), HttpStatus.CREATED);
 
         } catch (IOException e) {
-            System.err.println(e);
+            log.error("e: ", e);
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
     }
@@ -77,10 +87,10 @@ public class QuestionController {
             return new ResponseEntity<>(quizService.updateQuestion(validated), HttpStatus.OK);
 
         } catch (FileNotFoundException e) {
-            System.err.println(e);
+            log.error("e: ", e);
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         } catch (IOException e) {
-            System.err.println(e);
+            log.error("e: ", e);
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
     }
@@ -88,11 +98,18 @@ public class QuestionController {
     //Soft Delete
     @DeleteMapping(value = "/question")
     public ResponseEntity<String> deleteSoftQuestion(@RequestBody String uid) {
-        try{
-            quizService.deleteSoftQuestion(uid);
-            return new ResponseEntity<>("Question Removed", HttpStatus.OK);
+        try {
+            if (StringUtils.isAlphanumeric(uid)) {
+                quizService.softDeleteQuestion(uid);
+                return new ResponseEntity<>("Question Removed", HttpStatus.OK);
+            } else {
+                throw new IOException("Invalid Uid. Retry with valid uid.");
+            }
         } catch (FileNotFoundException e) {
-            System.err.println(e);
+            log.error("e: ", e);
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        } catch (IOException e) {
+            log.error("e: ", e);
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
     }
@@ -101,10 +118,18 @@ public class QuestionController {
     @DeleteMapping(value = "/delete")
     public ResponseEntity<String> deleteQuestion(@RequestBody String uid) {
         try{
-            quizService.deleteQuestion(uid);
-            return new ResponseEntity<>("Question Removed", HttpStatus.OK);
+            if (StringUtils.isAlphanumeric(uid)) {
+                quizService.deleteQuestion(uid);
+                return new ResponseEntity<>("Question Removed", HttpStatus.OK);
+            } else {
+                throw new IOException("Invalid Uid. Retry with valid uid.");
+            }
+
         } catch (FileNotFoundException e) {
-            System.err.println(e);
+            log.error("e: ", e);
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }  catch (IOException e) {
+            log.error("e: ", e);
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
     }
